封装蓝牙消息处理功能
